Change files in preparation for new CommandBox
