+++ src/main/java/chopchop/ui/CommandBox.java
@@ -1,0 +1,80 @@
+package chopchop.ui;
+;
+import chopchop.logic.commands.exceptions.CommandException;
+import chopchop.logic.parser.exceptions.ParseException;
+import javafx.collections.ObservableList;
+import javafx.fxml.FXML;
+import javafx.scene.control.TextField;
+import javafx.scene.layout.Region;
+
+/**
+ * The UI component that is responsible for receiving user command inputs.
+ */
+public class CommandBox extends UiPart<Region> {
+
+    public static final String ERROR_STYLE_CLASS = "error";
+    private static final String FXML = "CommandBox2.fxml";
+
+    private final CommandExecutor commandExecutor;
+
+    @FXML
+    private TextField commandTextField;
+
+    /**
+     * Constructs {@code CommandBox}
+     */
+    public CommandBox(CommandExecutor commandExecutor) {
+        super(FXML);
+        this.commandExecutor = commandExecutor;
+        // calls #setStyleToDefault() whenever there is a change to the text of the command box.
+        commandTextField.textProperty().addListener((unused1, unused2, unused3) -> setStyleToDefault());
+    }
+
+    /**
+     * Handles the Enter button pressed event.
+     */
+    @FXML
+    private void handleCommandEntered() {
+        try {
+            commandExecutor.execute(commandTextField.getText());
+        } catch (CommandException | ParseException e) {
+            setStyleToIndicateCommandFailure();
+        } finally {
+            commandTextField.setText("");
+        }
+    }
+
+    /**
+     * Sets the command box style to use the default style.
+     */
+    private void setStyleToDefault() {
+        commandTextField.getStyleClass().remove(ERROR_STYLE_CLASS);
+    }
+
+    /**
+     * Sets the command box style to indicate a failed command.
+     */
+    private void setStyleToIndicateCommandFailure() {
+        ObservableList<String> styleClass = commandTextField.getStyleClass();
+
+        if (styleClass.contains(ERROR_STYLE_CLASS)) {
+            return;
+        }
+
+        styleClass.add(ERROR_STYLE_CLASS);
+    }
+
+    /**
+     * Represents a function that can execute commands.
+     */
+    @FunctionalInterface
+    public interface CommandExecutor {
+        /**
+         * Executes the command and returns the result.
+         *
+         * @see chopchop.logic.Logic#execute(String)
+         */
+        chopchop.logic.commands.CommandResult execute(String commandText) throws CommandException, ParseException;
+    }
+
+}
