+++ src/main/java/chopchop/ui/CommandOutput.java
@@ -2,27 +2,17 @@ package chopchop.ui;
 
 import static java.util.Objects.requireNonNull;
 
-import chopchop.logic.commands.CommandResult;
-import chopchop.logic.commands.exceptions.CommandException;
-import chopchop.logic.parser.exceptions.ParseException;
-import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextArea;
-import javafx.scene.control.TextField;
 import javafx.scene.layout.Region;
 
 /**
  * The UI component that is responsible for receiving user command inputs.
  */
-public class CommandBox extends UiPart<Region> {
+public class CommandOutput extends UiPart<Region> {
 
-    public static final String ERROR_STYLE_CLASS = "error";
-    private static final String FXML = "CommandBox2.fxml";
+    private static final String FXML = "CommandOutput.fxml";
 
-    private final CommandExecutor commandExecutor;
-
-    @FXML
-    private TextField commandTextField;
 
     @FXML
     private TextArea displayBox;
@@ -30,45 +20,8 @@ public class CommandBox extends UiPart<Region> {
     /**
      * Constructs {@code CommandBox}
      */
-    public CommandBox(CommandExecutor commandExecutor) {
+    public CommandOutput() {
         super(FXML);
-        this.commandExecutor = commandExecutor;
-        // calls #setStyleToDefault() whenever there is a change to the text of the command box.
-        commandTextField.textProperty().addListener((unused1, unused2, unused3) -> setStyleToDefault());
-    }
-
-    /**
-     * Handles the Enter button pressed event.
-     */
-    @FXML
-    private void handleCommandEntered() {
-        try {
-            commandExecutor.execute(commandTextField.getText());
-        } catch (CommandException | ParseException e) {
-            setStyleToIndicateCommandFailure();
-        } finally {
-            commandTextField.setText("");
-        }
-    }
-
-    /**
-     * Sets the command box style to use the default style.
-     */
-    private void setStyleToDefault() {
-        commandTextField.getStyleClass().remove(ERROR_STYLE_CLASS);
-    }
-
-    /**
-     * Sets the command box style to indicate a failed command.
-     */
-    private void setStyleToIndicateCommandFailure() {
-        ObservableList<String> styleClass = commandTextField.getStyleClass();
-
-        if (styleClass.contains(ERROR_STYLE_CLASS)) {
-            return;
-        }
-
-        styleClass.add(ERROR_STYLE_CLASS);
     }
 
     /**
@@ -80,18 +33,4 @@ public class CommandBox extends UiPart<Region> {
         requireNonNull(feedbackToUser);
         displayBox.setText(feedbackToUser);
     }
-
-    /**
-     * Represents a function that can execute commands.
-     */
-    @FunctionalInterface
-    public interface CommandExecutor {
-        /**
-         * Executes the command and returns the result.
-         *
-         * @see seedu.address.logic.Logic#execute(String)
-         */
-        CommandResult execute(String commandText) throws CommandException, ParseException;
-    }
-
 }
