+++ src/main/resources/view/CommandOutput.fxml
@@ -15,7 +15,7 @@
          <children>
             <StackPane alignment="CENTER_LEFT">
                <children>
-                  <TextField fx:id="commandTextField" maxHeight="48.0" maxWidth="290.0" onAction="#handleCommandEntered" promptText="Enter command here..." />
+                  <TextField fx:id="commandTextField" maxHeight="48.0" maxWidth="290.0" promptText="Enter command here..." />
                </children>
             </StackPane>
             <TextArea fx:id="displayBox" prefHeight="208.0" prefWidth="304.0" text="" wrapText="true" editable="false"/>
