+++ src/main/resources/view/DisplayPanel.fxml
@@ -6,7 +6,7 @@
 
 <BorderPane prefHeight="520.0" prefWidth="624.0" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1">
     <center>
-        <StackPane fx:id="displayAreaPlaceholder" prefHeight="470.0" prefWidth="624.0" BorderPane.alignment="TOP_CENTER" />
+        <StackPane fx:id="displayAreaPlaceholder" prefHeight="422.0" prefWidth="624.0" BorderPane.alignment="TOP_CENTER" />
     </center>
    <bottom>
         <HBox fillHeight="false" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="59.0" prefWidth="624.0" BorderPane.alignment="TOP_CENTER">
@@ -74,4 +74,7 @@
             </children>
         </HBox>
    </bottom>
+   <top>
+       <TextField fx:id="commandTextField" maxHeight="48.0" maxWidth="600.0" promptText="Enter command here..." BorderPane.alignment="BOTTOM_CENTER"/>
+   </top>
 </BorderPane>
