+++ src/main/resources/view/IngredientCard.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.*?>
 
 <StackPane xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1">
-    <Button mnemonicParsing="false" prefHeight="140.0" prefWidth="168.0" styleClass="recipe-button">
+    <Button mnemonicParsing="false" prefHeight="125.0" prefWidth="148.0" styleClass="recipe-button">
         <graphic>
             <Label fx:id="name" styleClass="recipe_name_label" text="\$name" wrapText="true" />
         </graphic></Button>
