+++ src/main/resources/view/IngredientViewPanel.fxml
@@ -6,11 +6,11 @@
 
 <BorderPane xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1">
     <bottom>
-        <ScrollPane fx:id="ingredientPanel" pannable="true" prefHeight="470.0" prefWidth="624.0" vbarPolicy="NEVER" BorderPane.alignment="CENTER">
+        <ScrollPane fx:id="ingredientPanel" pannable="true" prefHeight="422.0" prefWidth="624.0" vbarPolicy="NEVER" BorderPane.alignment="TOP_CENTER">
             <content>
                 <GridPane fx:id="ingredientGridView" gridLinesVisible="true" hgap="20" vgap="10">
                     <padding>
-                        <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
+                        <Insets bottom="20.0" left="20.0" right="20.0" top="0.0" />
                     </padding></GridPane>
             </content>
         </ScrollPane>
