+++ src/main/resources/view/RecipeCard.fxml
@@ -9,11 +9,11 @@
    <stylesheets>
       <URL value="@Style.css" />
    </stylesheets>
-   <Button mnemonicParsing="false" onAction="#handleSelectRecipe" prefHeight="140.0" prefWidth="168.0" styleClass="recipe-button">
+   <Button mnemonicParsing="false" onAction="#handleSelectRecipe" prefHeight="125.0" prefWidth="148.0" styleClass="recipe-button">
       <graphic>
          <VBox alignment="CENTER">
             <children>
-               <ImageView fitHeight="70.0" fitWidth="90.0" pickOnBounds="true" preserveRatio="true" >
+               <ImageView fitHeight="60.0" fitWidth="80.0" pickOnBounds="true" preserveRatio="true" >
                   <image>
                      <Image url="@../images/recipe.png" />
                   </image>
@@ -21,15 +21,15 @@
                <Label fx:id="name" alignment="CENTER" styleClass="recipe_name_label" text="\$name" wrapText="true" />
                <HBox prefHeight="100.0" prefWidth="200.0" alignment="CENTER">
                   <children>
-                     <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
+                     <ImageView fitHeight="16.0" fitWidth="16.0" pickOnBounds="true" preserveRatio="true">
                         <image>
                            <Image url="@../images/pink_timer.png" />
                         </image>
                      </ImageView>
                      <Pane prefHeight="4.0" prefWidth="4.0"></Pane>
                      <Label text="X mins" />
-                     <Pane prefHeight="10.0" prefWidth="20.0"></Pane>
-                     <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
+                     <Pane prefHeight="5.0" prefWidth="20.0"></Pane>
+                     <ImageView fitHeight="16.0" fitWidth="16.0" pickOnBounds="true" preserveRatio="true">
                         <image>
                            <Image url="@../images/pax.png" />
                         </image>
